+++ kernel/signal.h
@@ -17,6 +17,7 @@ struct task_struct;
 #define SIGKILL   9
 #define SIGTERM   15
 #define SIGCHLD   17
+#define SIGSTOP   19
 #define NSIG      32
 
 /* Signal actions */